Add endpoints for changing player score and paying tax
Update models to use float for score changes and add reason field

+++ src/db_models.py
@@ -67,7 +67,7 @@ class PlayerScoreChange(DbBase):
         Integer, default=utc_now_ts, onupdate=utc_now_ts
     )
     player_id: Mapped[int] = mapped_column(Integer, nullable=False)
-    score_change: Mapped[int] = mapped_column(Integer, nullable=False)
+    score_change: Mapped[int] = mapped_column(Float, nullable=False)
     reason: Mapped[str] = mapped_column(String(255), nullable=False)
     sector_id: Mapped[int] = mapped_column(Integer, nullable=False)
